Refactor CourseInfoPopup into separate component
