+++ pages/course/[id].tsx
@@ -14,6 +14,10 @@ import { server } from '../../lib/server';
 import { MdChecklist } from 'react-icons/md';
 import { BsPerson } from 'react-icons/bs';
 import ExpandableText from '../../components/expandableText';
+import {
+  CourseInfoPopupObject,
+  TransitionWrapper,
+} from '../../components/courseInfoPopup';
 
 const CoursePage = ({
   params,
@@ -92,116 +96,22 @@ const CoursePage = ({
     }
   }, [graph, currentlyHoveredId]);
 
-  interface CourseInfoPopupParams {
-    active: boolean; // whether the popup is currently active
-    longTitle: string;
-    course_no: string;
-    desc: string;
-    eli: string;
-    prereqFull: string;
-    locked: boolean;
-  }
-  // Mouse coordinates - determines where to display popup
-  const [coords, setCoords] = useState({ x: 0, y: 0 });
-  // Parameters for course info popup (opened when mouse hovers over node)
-  const getEmptyPopupParams = () => {
-    return {
-      active: false,
-      longTitle: '',
-      course_no: '',
-      desc: '',
-      eli: '',
-      prereqFull: '',
-      locked: false,
-    } as CourseInfoPopupParams;
-  };
-  const initialPopupParams = getEmptyPopupParams();
-  const [courseInfoPopupParams, setCourseInfoPopupParams] =
-    useState<CourseInfoPopupParams>(initialPopupParams);
-  // Track mouse position
-  useEffect(() => {
-    const handleMouseMove = (e: MouseEvent) => {
-      if (courseInfoPopupParams.locked) return;
-      setCoords({ x: e.pageX, y: e.pageY });
-    };
-    window.addEventListener('mousemove', handleMouseMove);
-    return () => {
-      window.removeEventListener('mousemove', handleMouseMove);
-    };
-  }, [coords, courseInfoPopupParams.locked]);
-
-  function CourseInfoPopup() {
-    const cipp = courseInfoPopupParams;
-    return (
-      <div
-        className="m-5 max-w-lg rounded-lg bg-gray-900/80 text-white backdrop-blur"
-        style={{
-          display: cipp.active ? 'block' : 'none',
-          position: 'absolute',
-          left: coords.x, // If there isn't enough margin/offset, nodeUnhoverCallback will trigger once this opens because the cursor will be over this popup instead of the node
-          top: coords.y - 35,
-          transform: 'translate(0, -100%)',
-          zIndex: 100,
-        }}
-      >
-        <p className="ml-2 mr-2 mt-2 text-xl font-bold">
-          {cipp.longTitle}{' '}
-          {cipp.course_no != 'PEA000' ? ' · ' + cipp.course_no : ''}
-        </p>
-        <p className="ml-2 mr-2 text-sm">{cipp.desc}</p>
-        <p className="ml-2 mr-2 text-sm italic">{cipp.eli}</p>
-        <p className="ml-2 mr-2 mb-2 text-sm italic">
-          {cipp.prereqFull == '' ? '' : `Prerequisite(s): ${cipp.prereqFull}`}
-        </p>
-      </div>
-    );
-  }
-  // Callbacks for when user moves cursors on/off nodes or clicks on nodes
-  interface flowNode {
-    id: string;
-  }
-
-  const nodeHoverCallback = (event: React.MouseEvent, node: flowNode) => {
-    setCurrentlyHoveredId(node.id);
-    if (courseInfoPopupParams.locked) return;
-    const popupParams = {
-      active: true,
-      longTitle: titles[node.id],
-      course_no: node.id,
-      desc: descriptions[node.id],
-      eli: eli[node.id],
-      prereqFull: prereqFull[node.id],
-      locked: false,
-    } as CourseInfoPopupParams;
-
-    setCourseInfoPopupParams(popupParams);
-  };
-  const nodeUnhoverCallback = () => {
-    setCurrentlyHoveredId('');
-    if (courseInfoPopupParams.locked) return;
-    setCourseInfoPopupParams(getEmptyPopupParams());
-  };
-  // Lock/unlock course info popup when right click on popup
-  const nodeRightClickCallback = (event: React.MouseEvent) => {
-    event.preventDefault();
-    const popupParams = courseInfoPopupParams;
-    popupParams.locked = !popupParams.locked;
-    setCourseInfoPopupParams(popupParams);
-  };
-  // Unlock course info popup when click on canvas
-  const paneClickCallback = () => {
-    console.log('pane clicked');
-    setCourseInfoPopupParams(getEmptyPopupParams());
-  };
-  // Open the course page associated with this course
-  const nodeClickCallback = (event: React.MouseEvent, element: flowNode) => {
-    event.preventDefault();
-    // check if element is an edge
-    if (element.id.startsWith('pe') || element.id.startsWith('ce')) return;
-    if (event.metaKey || event.ctrlKey) {
-      window.open(`/course/${element.id}`);
-    } else window.open(`/course/${element.id}`, '_self');
-  };
+  const {
+    nodeHoverCallback,
+    nodeUnhoverCallback,
+    nodeRightClickCallback,
+    paneClickCallback,
+    nodeClickCallback,
+    courseInfoPopupParams,
+    coords,
+    CourseInfoPopup,
+  } = CourseInfoPopupObject({
+    setCurrentlyHoveredId,
+    titles,
+    descriptions,
+    eli,
+    prereqFull,
+  });
 
   // Advances to the next level of requirements - called when "More Prereqs" is clicked
   const getMoreReqs = async () => {
@@ -367,7 +277,12 @@ const CoursePage = ({
               <Background color="#858585" />
             </ReactFlow>
           </div>
-          <CourseInfoPopup />
+          <TransitionWrapper
+            courseInfoPopupParams={courseInfoPopupParams}
+            coords={coords}
+          >
+            <CourseInfoPopup />
+          </TransitionWrapper>
         </div>
       </div>
     </div>
